only run gpg-agent stuff on local machines

+++ bashrc
@@ -35,6 +35,7 @@ bind '"\e.":yank-last-arg'
 
 # local hostnames for my machines to set local PS1 colorscheme vs remote
 _localhosts=("pseudogen" "stream")
+short_hostname=${HOSTNAME%%.*}
 
 export EDITOR=vim
 
@@ -63,26 +64,28 @@ if rpm -q vim-common &> /dev/null; then
     alias vless="$(rpm -ql vim-common | grep less.sh)"
 fi
 
-# Ensure gpg-agent starts with --enable-ssh-support
-if [[ $EUID -ne 0 ]]; then
-    if [[ ! -f /run/user/$(id -u)/gpg-agent.env ]]; then
-        if [[ -f /usr/bin/gpg-agent ]]; then
-            killall gpg-agent &> /dev/null;
-            eval $( \
-                gpg-agent \
-                    --daemon \
-                    --enable-ssh-support \
-                    > /run/user/$(id -u)/gpg-agent.env
-            )
+# Ensure gpg-agent starts with --enable-ssh-support when on local machines
+if [[ ${_localhosts[@]} =~ ${short_hostname} ]]; then
+    if [[ $EUID -ne 0 ]]; then
+        if [[ ! -f /run/user/$(id -u)/gpg-agent.env ]]; then
+            if [[ -f /usr/bin/gpg-agent ]]; then
+                killall gpg-agent &> /dev/null;
+                eval $( \
+                    gpg-agent \
+                        --daemon \
+                        --enable-ssh-support \
+                        > /run/user/$(id -u)/gpg-agent.env
+                )
+            fi
         fi
     fi
-fi
-if [[ -f /run/user/$(id -u)/gpg-agent.env ]]; then
-    . /run/user/$(id -u)/gpg-agent.env
-fi
-# Make sure gnome keyring doesn't do anything with ssh keys
-if [[ $(gconftool-2 --get /apps/gnome-keyring/daemon-components/ssh) != "false" ]]; then
-    gconftool-2 --type bool --set /apps/gnome-keyring/daemon-components/ssh false
+    if [[ -f /run/user/$(id -u)/gpg-agent.env ]]; then
+        . /run/user/$(id -u)/gpg-agent.env
+    fi
+    # Make sure gnome keyring doesn't do anything with ssh keys
+    if [[ $(gconftool-2 --get /apps/gnome-keyring/daemon-components/ssh) != "false" ]]; then
+        gconftool-2 --type bool --set /apps/gnome-keyring/daemon-components/ssh false
+    fi
 fi
 
 
@@ -265,8 +268,6 @@ __my_vcs_prompt () {
 # 'set show-mode-in-prompt on' (requires bash 4.3+ and readline 6.3+)
 __prompt_command() {
     local exit_code=$?
-
-    local short_hostname=${HOSTNAME%%.*}
 
     local prompt_out=""
